refactor customizations & packages; λαω package; show modeline time
Refactor Emacs customizations. Consolidate customizations in a single
Emacs use-package macro.

Move theme customizations and loading from early init to init.

Set package-user-dir to be in the Emacs user "var" directory.

Consolidate λαω files under one package and load it in init file.

Move most functions defined in the init file to λαω-functions. This
includes:

1) a function to display Emacs init load time as a message
2) a function to remove kill ring text properties upon quitting Emacs
3) a function for showing separators in the completing-read-multiple
prompt indicator
4) a function to truncate lines locally

Show the time in the modeline.

Set :ensure to nil for some built-in packages.